[logiclayer] refactors helper methods
