+++ tesseract-core/Cargo.toml
@@ -10,6 +10,7 @@ csv = "1.0.2"
 failure = "0.1.3"
 futures = "0.1.25"
 itertools = "0.7.11"
+log = "0.4.3"
 serde = "1.0.71"
 serde-xml-rs = "0.3.1"
 serde_derive = "1.0.71"
