+++ tesseract-server/src/logic_layer/mod.rs
@@ -4,4 +4,4 @@ mod util;
 
 pub use self::cache::{Cache, CubeCache, Time, TimePrecision, TimeValue, populate_cache};
 pub use self::config::{LogicLayerConfig, read_config};
-pub use self::util::{stringify_column_data, boxed_error};
+pub use self::util::{boxed_error};
